Fix stride computation formula used during compute estimation
Turns out the previous PR
https://github.com/pytorch-labs/autoparallel/pull/37

was not correct. It divided the wrong dim's stride.

This PR divides the dim to the left of the one being sharded, which is
what really happens.

Note: that we have this util at all is worrying me. Why don't we just
use dtensors to propagate?