Use custom redistribute_cost for optimal PP->S(0)S(0) cost
Previously, given the default iteration order, it was always less expensive to do PP->S(0)P->S(0)S(0) instead of directly PP->S(0)S(0). We now favor doing it in a single pass by uting the optimal redistribution cost for a given operation. For now, we only consider the PP->S(0)S(0) case, but we should generalize this to all cases in the future