Fix clustering key to account for input node placements as well
Previously, we were only taking the current node OpSpec string to hash it. This includes the supported input/output placements, but doesn't account for its input node, which might have different output shardings. This PR fixes it, and also add an assert in the optimize_sharding to validate that the created links are consistent