Remove stack decomposition and add stack rule
Using the stack decomposition leads to fewer sharding options. As an example, two S(0) tensors can be successfully stacked together at dimension 0, which should lead to a S(1) output sharding. If we instead keep the decomposition from https://github.com/pytorch/pytorch/blob/ded9bcd61a059bf723e6e84689552962b480ea77/torch/_refs/__init__.py#L4116, which first concatenates at the stack dim and then applies a view, we can't obtain the same sharding option. This is because stack has a stricter set of requirements as cat, which the decomposition makes us miss it. Once I removed the decomposition, I faced an issue that the propagation rules from stack aren't correctly implemented, so I had to re-implement it. I'm following a much simpler pattern for the propagation rules, which is to enumerate all possible sharding options and expand the mesh afterwards, which makes the implementation much simpler. This I believe is in-line with what @wconstab is doing for his refactoring of the propagation rules